+++ RELEASENOTES.md
@@ -3,6 +3,8 @@
 ### Unreleased changes
 
 *   Common Library:
+    *   Remove `Format.toBundle(boolean excludeMetadata)` method, use
+        `Format.toBundle()` instead.
 *   ExoPlayer:
     *   Consider language when selecting a video track. By default select a
         'main' video track that matches the language of the selected audio